address some safari issues
Safari has some bugs with loading=lazy images, and also does some
strange things related to the `allow-discrete` transition stuff.

The symptom with the loading=lazy images is that if there are a bunch of
images in the group loaded in the lightbox and those images are not yet
actually loaded, Safari seems to calculate their size *before* actually
fetching the images, and not re-flow after fetching. The result is those
images in the thumbnails bar end up overlapping each other since they
are calculated as having very low widths before the image is fetched but
then actually rendered at the appropriate width after the image is
fetched.

Fix for that is I'm abandoning the attempt to keep thumbnail aspect
ratio accurate to the original image and I'm just forcing everything to
be square.

The bug with the transition stuff was that closing the modal would
immediately disappear the <dialog> itself (the background would go
away), but contents (the image, buttons, description <p>) would remain
on screen for a moment and then disappear. Fix for that for now is to
just... remove the `transition:` property on the dialog. I'm commenting
it out instead of deleting, I'd like to come back to this at some point.

Really is delightful how Safari is basically the modern IE: it lags its
competition in implementing standards and when it does implement them
they're often semi-broken.